py/requirements.txt
pika
jieba
juba
cassandra-driver
schedule
requests